refactor(signer)!: move around
- move `GenericProxySigner` from common to manager
- simplify inners of `EcdsaSignature`
