+++ crates/signer/src/manager.rs
@@ -4,14 +4,70 @@ use alloy::rpc::types::beacon::{BlsPublicKey, BlsSignature};
 use blst::min_pk::SecretKey as BlsSecretKey;
 use cb_common::{
     commit::request::{ProxyDelegation, SignedProxyDelegation},
-    signer::{
-        EcdsaSecretKey, GenericProxySigner, GenericPubkey, ProxySigner, PubKey, SecretKey, Signer,
-    },
+    signer::{EcdsaSecretKey, GenericPubkey, PubKey, SecretKey, Signer},
     types::{Chain, ModuleId},
 };
+use derive_more::derive::{Deref, From};
 use tree_hash::TreeHash;
 
 use crate::error::SignerModuleError;
+
+// For extra safety and to avoid risking signing malicious messages, use a proxy
+// setup: proposer creates a new ephemeral keypair which will be used to sign
+// commit messages, it also signs a ProxyDelegation associating the new keypair
+// with its consensus pubkey When a new commit module starts, pass the
+// ProxyDelegation msg and then sign all future commit messages with the proxy
+// key for slashing the faulty message + proxy delegation can be used
+// Signed using builder domain
+#[derive(Clone, Deref)]
+pub struct ProxySigner<T: SecretKey> {
+    #[deref]
+    signer: Signer<T>,
+    delegation: SignedProxyDelegation,
+}
+
+impl<T: SecretKey> ProxySigner<T> {
+    pub fn new(signer: Signer<T>, delegation: SignedProxyDelegation) -> Self {
+        Self { signer, delegation }
+    }
+}
+
+#[derive(From)]
+pub enum GenericProxySigner {
+    Bls(ProxySigner<BlsSecretKey>),
+    Ecdsa(ProxySigner<EcdsaSecretKey>),
+}
+
+impl GenericProxySigner {
+    pub fn pubkey(&self) -> GenericPubkey {
+        match self {
+            GenericProxySigner::Bls(proxy_signer) => GenericPubkey::Bls(proxy_signer.pubkey()),
+            GenericProxySigner::Ecdsa(proxy_signer) => GenericPubkey::Ecdsa(proxy_signer.pubkey()),
+        }
+    }
+
+    pub fn delegation(&self) -> SignedProxyDelegation {
+        match self {
+            GenericProxySigner::Bls(proxy_signer) => proxy_signer.delegation,
+            GenericProxySigner::Ecdsa(proxy_signer) => proxy_signer.delegation,
+        }
+    }
+
+    pub async fn sign(&self, chain: Chain, object_root: [u8; 32]) -> Vec<u8> {
+        match self {
+            GenericProxySigner::Bls(proxy_signer) => {
+                proxy_signer.sign(chain, object_root).await.to_vec()
+            }
+            GenericProxySigner::Ecdsa(proxy_signer) => {
+                proxy_signer.sign(chain, object_root).await.to_vec()
+            }
+        }
+    }
+
+    pub async fn verify_signature(&self, msg: &[u8], signature: &[u8]) -> eyre::Result<()> {
+        self.pubkey().verify_signature(msg, signature)
+    }
+}
 
 #[derive(Default)]
 struct ProxySigners {
